requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
dj-database-url==3.0.0
Django==5.2.3
djangorest-framework==3.16.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.0
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.2.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
uvicorn==0.34.3
